// src/app/api/models/effective-settings-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Response containing effective settings for a user
 */
export interface EffectiveSettingsResponse {
  settings?: {
[key: string]: {
[key: string]: any;
};
};
  tenantId?: string | null;
  userId?: string;
}